+++ packages/react/src/atomic/Button/src/components/ButtonAnchor/index.ts
@@ -1,0 +1,4 @@
+import { ButtonAnchor } from "./ButtonAnchor";
+
+export { ButtonAnchor } from "./ButtonAnchor";
+export default ButtonAnchor;